# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API bitkakir project",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/users/forget-recover-password": {
            "post": {
                "tags": [
                    "Forget Password"
                ],
                "summary": "Recoved Password",
                "description": "Realization recoved password, inform you phone number, password and confirm password",
                "operationId": "b6d4d51bec9285c6fa4378631da1c390",
                "requestBody": {
                    "description": "Realization recoved password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phoneNumber",
                                    "passowrd",
                                    "confirm"
                                ],
                                "properties": {
                                    "phoneNumber": {
                                        "description": "You phone number",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Inform you password",
                                        "type": "string"
                                    },
                                    "confirm": {
                                        "description": "Confirm you password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Forget password successful"
                    }
                }
            }
        },
...
        "/api/v1/user-types": {
            "get": {
                "tags": [
                    "User types"
                ],
                "summary": "List user types",
                "description": "Display listing of the resource",
                "operationId": "e2036f3d141606ea6cdf2a6f3e0f2ca4",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "User types"
                ],
                "summary": "Save user type",
                "description": "Creates and returns a new  user type",
                "operationId": "1f6826f669f2a3421bc2ca0d7ce0979d",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ServiceRequest"
                },
                "responses": {
                    "201": {
                        "description": "Service created successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            }
        },
...
        "/api/v1/sms-opt": {
            "post": {
                "tags": [
                    "Send OTP"
                ],
                "summary": "Get code otp for create account",
                "description": "Send code OTP in phone number",
                "operationId": "b7370788a9113a974fc0e73820757d1e",
                "requestBody": {
                    "description": "Send OTP for create accout",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phoneNumber"
                                ],
                                "properties": {
                                    "phoneNumber": {
                                        "description": "Your phone number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    }
                }
            }
        },
        "/api/v1/sms-opt-forget": {
            "post": {
                "tags": [
                    "Send OTP"
                ],
                "summary": "Get code otp for recoved password",
                "description": "Send code OTP in phone number",
                "operationId": "cffd4185bfab3d6659911e8ada27665b",
                "requestBody": {
                    "description": "Send OTP for recoved password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phoneNumber"
                                ],
                                "properties": {
                                    "phoneNumber": {
                                        "description": "Your phone number",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    }
                }
            }
        },
...
        "/api/v1/user-types/{id}": {
            "get": {
                "tags": [
                    "User types"
                ],
                "summary": "Show  user type by UUID or ID",
                "description": "Display a listing by UUID or ID",
                "operationId": "ed64de1f0198dc2a06dc4a5979b06f9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search by UUID or ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "User types"
                ],
                "summary": "Update user type",
                "description": "Update the specified resource in storage",
                "operationId": "2276575b221311880b8da2d303e79659",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Search by UUID or ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/ServiceRequest"
                },
                "responses": {
                    "201": {
                        "description": "Service updated successfully"
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User types"
                ],
                "summary": "Remove the specified resource from storage",
                "description": "Delete user type",
                "operationId": "2383dddb0429e245ed02d0fbe8355d11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Delete by UUID or ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "requestBodies": {
...
            "ServiceRequest": {
                "description": "Dados do serviço",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Consultoria em Marketing"
                                },
                                "image": {
                                    "type": "string",
                                    "example": "servico.jpg"
                                },
                                "video": {
                                    "type": "string",
                                    "example": "apresentacao.mp4"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Descrição completa do serviço."
                                },
                                "user_id": {
                                    "type": "string",
                                    "example": "1"
                                },
                                "category_id": {
                                    "type": "string",
                                    "example": "2"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        }
    },
...
}